Rename FeedViewController to ListViewController since it can now render any collection of cell controllers
