+++ EssentialApp/EssentialAppTests/FeedAcceptanceTests.swift
@@ -60,13 +60,13 @@ class FeedAcceptanceTests: XCTestCase {
     private func launch(
         httpClient: HTTPClientStub = .offline,
         store: InMemoryFeedStore = .empty
-    ) -> FeedViewController {
+    ) -> ListViewController {
         let sut = SceneDelegate(httpClient: httpClient, store: store)
         sut.window = UIWindow()
         sut.configureWindow()
 
         let nav = sut.window?.rootViewController as? UINavigationController
-        return nav?.topViewController as! FeedViewController
+        return nav?.topViewController as! ListViewController
     }
 
     private func enterBackground(with store: InMemoryFeedStore) {
